+++ Editor/Scripts/Menu Editor Visualizer/EditorMenuEditorVisualizer.cs
@@ -17,32 +17,25 @@ namespace SLIDDES.UI.MenuEditorVisualizer
 
         public override void OnInspectorGUI()
         {
+            GUILayout.BeginHorizontal();
+            if(GUILayout.Button("Organize Menus", GUILayout.Height(32)))
+            {
+                selected.OrganiseMenus();
+            }
+            if(GUILayout.Button("Reset Menus", GUILayout.Height(32)))
+            {
+                selected.ResetMenus();
+            }
+            GUILayout.EndHorizontal();
+
+            selected.logDebugs = GUILayout.Toggle(selected.logDebugs, new GUIContent("Log Debugs", "Log the debug.logs of this script to the console"));
+            selected.useCustomRectTransforms = GUILayout.Toggle(selected.useCustomRectTransforms, new GUIContent("Use Custom Rect Transforms", "Do you want to set your own RectTransforms instead of the script auto assigning them?"));
             if(selected.useCustomRectTransforms)
             {
-                if(GUILayout.Button("Organize Menus", GUILayout.Height(32)))
-                {
-                    selected.OrganiseMenus();
-                }
-                GUILayout.Space(8);
-                if(GUILayout.Button("Reset Menus", GUILayout.Height(32)))
-                {
-                    selected.ResetMenus();
-                }
-                base.OnInspectorGUI();
-            }
-            else
-            {
-                if(GUILayout.Button("Organize Menus", GUILayout.Height(32)))
-                {
-                    selected.OrganiseMenus();
-                }
-                GUILayout.Space(8);
-                if(GUILayout.Button("Reset Menus", GUILayout.Height(32)))
-                {
-                    selected.ResetMenus();
-                }
-                GUILayout.Space(8);
-                selected.useCustomRectTransforms = GUILayout.Toggle(selected.useCustomRectTransforms, new GUIContent("Use Custom Rect Transforms", "Do you want to set your own RectTransforms instead of the script auto assigning them?"));
+                SerializedObject so = new SerializedObject(selected);
+                SerializedProperty sp = so.FindProperty("menus");
+                EditorGUILayout.PropertyField(sp, true);
+                so.ApplyModifiedProperties();
             }
         }
     }