+++ Runtime/Scripts/UI Navigator/UINavigatorMenu.cs
@@ -12,26 +12,35 @@ namespace SLIDDES.UI.Navigator
     [AddComponentMenu("SLIDDES/UI/UI Navigator Menu")]
     public class UINavigatorMenu : MonoBehaviour
     {
-        [Tooltip("Dont close the menu on start")]
-        public bool dontCloseOnStart;
+        [Tooltip("The way this menu closes on application start")]
+        [SerializeField] public BeginClosingMethod beginClosingMethod;
         [Tooltip("Include the gameobject where the UINavigatorMenu is attached too to be turned on/off")]
-        [SerializeField] private bool includeSelf;
-        [Tooltip("The elements of the menu to toggle")]
-        public GameObject[] elements;
+        [SerializeField] private bool includeClosingSelf;
 
         [Header("Callbacks")]
-        public UnityEvent onMenuOpen;
-        public UnityEvent onMenuClose;
-
-        [Header("Extra Menus")]
+        [Tooltip("The elements of the menu to toggle")]
+        public GameObject[] elementsToToggle;
         [Tooltip("Extra menus to open when this menu opens (and close when this menu closes)")]
         public UINavigatorMenu[] menusToOpen;
         [Tooltip("Extra menus to close when this menu opens (and open when this menu closes")]
         public UINavigatorMenu[] menusToClose;
+        [Space()]
+        public UnityEvent onMenuOpen;
+        public UnityEvent onMenuClose;
+
+        /// <summary>
+        /// To check if it still needs to use its own closing method or that another script has called this one Close() / Open() method so it wont need to do its own closingmethod
+        /// </summary>
+        private bool doClosingMethod = true;
+
+        private void Awake()
+        {
+            if(doClosingMethod && beginClosingMethod == BeginClosingMethod.onAwake) Close();
+        }
 
         private void Start()
         {
-            if(!dontCloseOnStart) Close();
+            if(doClosingMethod && beginClosingMethod == BeginClosingMethod.onStart) Close();
         }
 
         /// <summary>
@@ -39,9 +48,11 @@ namespace SLIDDES.UI.Navigator
         /// </summary>
         public void Open()
         {
-            if(includeSelf) gameObject.SetActive(true);
+            doClosingMethod = false;
 
-            foreach(GameObject item in elements)
+            if(includeClosingSelf) gameObject.SetActive(true);
+
+            foreach(GameObject item in elementsToToggle)
             {
                 if(item != null) item.SetActive(true);
             }
@@ -64,9 +75,11 @@ namespace SLIDDES.UI.Navigator
         /// </summary>
         public void Close()
         {
-            if(includeSelf) gameObject.SetActive(false);
+            doClosingMethod = false;
 
-            foreach(GameObject item in elements)
+            if(includeClosingSelf) gameObject.SetActive(false);
+
+            foreach(GameObject item in elementsToToggle)
             {
                 if(item != null) item.SetActive(false);
             }
@@ -89,14 +102,25 @@ namespace SLIDDES.UI.Navigator
         /// </summary>
         public void CloseSelf(bool invokeOnMenuClose = false)
         {
-            if(includeSelf) gameObject.SetActive(false);
+            if(includeClosingSelf) gameObject.SetActive(false);
 
-            foreach(GameObject item in elements)
+            foreach(GameObject item in elementsToToggle)
             {
                 if(item != null) item.SetActive(false);
             }
 
             if(invokeOnMenuClose) onMenuClose?.Invoke();
+        }
+
+        /// <summary>
+        /// The way this menu closes on application start
+        /// </summary>
+        [System.Serializable]
+        public enum BeginClosingMethod
+        {
+            dontClose,
+            onAwake,
+            onStart
         }
     }
 }